refactor: change mypage on UI
- 전체적으로 UI 구성 리팩토링 작업
- layout 부분 일부 리팩토링으로 가독성 상승
- css 모듈에 있는 내용 전부 제거하고 tailwindcss로만 구성하여 layout 설정

obab-client #13
