--- src/css/my/page/mobile/myPage.mobile.module.css
@@ -1,47 +1,0 @@
-/* mypage 전체 화면 컴포넌트 설정 */
-.myPageComponent {
-  /*display: flex;*/
-  /*flex-direction: column;*/
-  /*flex-wrap: wrap;*/
-  max-width: 100%;
-  margin: 0 auto
-}
-
-
-/* mypage 윗쪽 컴포넌트 */
-.myPageUpComponent {
-  display: flex;
-  flex-direction: row;
-  width:100%;
-  min-height: 10vh;
-}
-
-
-/* mypage 왼쪽 컴포넌트 이미지 스타일링 */
-.myPageUpProfileImageComponent {
-  margin: 0 auto;
-  width:100%;
-  /*overflow:hidden;*/
-  height:10vh;
-  align-items: center;
-  justify-items: center;
-}
-
-
-/* mypage 이미지 */
-.myPageUpProfileImage {
-  border-image-width: 1px;
-  width:100%;
-  height:100%;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-
-
-
-/* mypage 오른쪽 컴포넌트 */
-.myPageDownComponent {
-  width: 100%;
-  min-height:50vh;
-}
